Add CodePanelContext and move entry selection into individual Entry components
In preparation to add links into and out of Map Explorer, map explorer needed
it's own routing url structure. Deep linking into the Map Explorer requires a
different url structure than the other table URLs (and future UI may also need
this ability). In order to support different URL structures, FileViewer can't
assume the URL structure for all of it's children and determine the selected
entry ID. Since it (the parent of DeoptTables and CodePanel) can't hold the
state of the selected entry, the state of the selected entry needs to move into
a context so that deep children can communicate across the tree to the CodePanel
about what lines/entries to highlight.

Further, since Map Explorer may want to highlight lines without entries and the
URL can't be relied on to determine what marker should be highlighted, markers
can no longer be highlighted based on the location.hash matching the marker id
attribute. So we had to modify how markers are highlighted. Since makers are
created imperatively and not by Preact, a list of created marker elements is
returned to the CodePanel element so it can imperatively update markers when
they are highlighted.

To optimize rendering, CodePanel context is separated into context instances:
one for state, and one for updating state. This distinction allows components
that only dispatch changes to not rerender when the state updates.

In summary, the URL holds the state of what deopt info panel (codes, deopts,
ics, maps) to show. That panels components Route and Component understand if the
URL contains the state of which entry (if any, if relevant) is selected. The
deopt info panel must then update the application (e.g. CodePanelContext) with
the selected entry.

Also, adds a format route definition interface to support generating URLs